fix(smart_shift): Ignore comma
It's becoming cumbersome to have to re-engage smart shift when using Vim
commands. These commands often begin with `:` which causes smart shift
to be disabled.

This change adds `:` to our ignore list so that we are able to press
less keys to trigger these commands.

This makes using smart shift less cumbersome and more intuitive.

Signed-off-by: Yomi Colledge <baphled@boodah.net>